refactor plugins architecture
